+++ packages/terra-date-picker/src/terra-dev-site/test/date-picker/DatePickerControlled.test.jsx
@@ -1,6 +1,10 @@
 import React from 'react';
 import Button from 'terra-button';
+import classNames from 'classnames/bind';
 import DatePicker from '../../../DatePicker';
+import styles from './common/DatePicker.test.module.scss';
+
+const cx = classNames.bind(styles);
 
 class DatePickerDefault extends React.Component {
   constructor(props) {
@@ -28,7 +32,7 @@ class DatePickerDefault extends React.Component {
 
   render() {
     return (
-      <div>
+      <div className={cx('content-wrapper')}>
         <DatePicker
           name="controlled-date-picker"
           value={this.state.date}