+++ packages/terra-date-picker/src/terra-dev-site/test/date-picker/DatePickerDefaultDateBeforeMinDateOnly.test.jsx
@@ -1,8 +1,12 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import DatePicker from '../../../DatePicker';
+import styles from './common/DatePicker.test.module.scss';
+
+const cx = classNames.bind(styles);
 
 const DatePickerDefaultDateBeforeMinDateOnly = () => (
-  <div>
+  <div className={cx('content-wrapper')}>
     <h3>minDate Only</h3>
     <DatePicker
       name="date-input"