+++ packages/terra-date-picker/src/terra-dev-site/test/date-picker/DatePickerDisabled.test.jsx
@@ -1,12 +1,18 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import DatePicker from '../../../DatePicker';
+import styles from './common/DatePicker.test.module.scss';
+
+const cx = classNames.bind(styles);
 
 const DatePickerDisabledDate = () => (
-  <DatePicker
-    name="date-input"
-    selectedDate="2017-04-01"
-    disabled
-  />
+  <div className={cx('content-wrapper')}>
+    <DatePicker
+      name="date-input"
+      selectedDate="2017-04-01"
+      disabled
+    />
+  </div>
 );
 
 export default DatePickerDisabledDate;