+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -6,6 +6,17 @@ Unreleased
 ### Added
 * Added new test to validate missing hour scenario 
 
+4.16.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+### Added
+* Added `isIncomplete` prop and stylings
+* Added `isInvalid` prop and stylings
+* Added `isInvalidMeridiem` prop 
+* Added `required` prop 
+
 4.15.0 - (October 3, 2019)
 ------------------
 ### Changed