+++ packages/terra-slide-group/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+4.12.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
 4.11.0 - (October 3, 2019)
 ------------------
 ### Changed