+++ packages/terra-time-input/CHANGELOG.md
@@ -6,6 +6,17 @@ Unreleased
 ### Fixed
 * Fixed time-input componentDidUpdate for updating time-input on re-entering the same missing hour in time-input.
 
+4.13.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+### Added
+* Added `isIncomplete` prop and stylings
+* Added `isInvalid` prop and stylings
+* Added `isInvalidMeridiem` prop 
+* Added `required` prop
+
 4.12.0 - (October 3, 2019)
 ------------------
 ### Changed